Move find_bin_edges into utils.dimensions
